// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface Location {
    country: string;
    city: string;
    state: string;
}
// ...
export interface ContactInfo {
    email: string;
    website: string;
    address: string;
    phone: string;
}
// ...
export interface HealthcareProfessionalRequest {
    id: string;
    status: RequestStatus;
    requester: Principal;
    contact: ContactInfo;
    name: string;
    role: string;
    experience: bigint;
    credentials: Array<ExternalBlob>;
    specialties: Array<string>;
    location: Location;
}
export enum DemoBookingStatus {
    pending = "pending",
    completed = "completed",
    contacted = "contacted"
}
export enum RequestStatus {
    pending = "pending",
    approved = "approved",
    rejected = "rejected"
}